+++ src/app/tab1/tab1.page.html
@@ -16,6 +16,11 @@
     </ion-card-content>
   </ion-card>
   <ion-button (click)='openForm()' expand="full">Add new emergency situation</ion-button>
+  <ion-button (click)='sendMessage()' expand="full">DANGERit</ion-button>
+  <ion-card-content>
+      <p>Youn current latitude: {{myLocation.latitude}}</p>
+      <p>Youn current longitude: {{myLocation.longitude}}</p>
+    </ion-card-content>
   <ion-list lines="none">
     <ion-list-header>
       <ion-label>Resources</ion-label>
